Added clock, updated appearance, fixed issues
Added a TextClock view to the top of the frame.

Bus times are now shown in minutes:seconds format instead of the number
of seconds. They are also now shown in a list instead of simply using
the .toString of the array list.

Made the background grey instead of the default light blue. Got rid of
the "Dummy Content" and "Dummy Button."

TODO: fix the issue where anything but the first <direction> tag is
ignored in parsing bus times.